pyproject.toml
[tool.poetry]
name = "edea_ms"
version = "0.1.0"
description = "EDeA Measurement Server"
authors = ["Elen Eisendle <ln@calcifer.ee>"]
license = "EUPLv1.2"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
ormar = { extras = ["orjson"], version = "^0.12.0" }
uvicorn = { extras = ["standard"], version = "^0.19.0" }
databases = "^0.6.1"
aiosqlite = "^0.17.0"
httpx = "^0.23.0"
alembic = "^1.8.1"
aiofiles = "^22.1.0"
Jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
trio = "^0.22.0"
mypy = "^0.982"
sqlalchemy-stubs = "^0.4"
pytest-alembic = "^0.8.4"

[tool.poetry.group.debug]
optional = true

[tool.poetry.group.debug.dependencies]
yappi = "^1.3.6"
py-spy = "^0.3.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlmypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
